week_1/requirements.txt
groq
openai
streamlit
python-dotenv